minor refactoring of schemas
